refactor(website): deduplicate stepper visuals and fix file casing
Extract shared AppPreviewShell, DrawEdge utility, and layout-styles.
Remove useCallback wrappers, merge duplicate styled components, and
rename .tsx files to PascalCase per project convention.
