+++ packages/twenty-website-new/src/sections/ProductStepper/visuals/DataModelVisual.tsx
@@ -1,10 +1,11 @@
 'use client';
 
 import { styled } from '@linaria/react';
-import { useCallback, useRef, useState } from 'react';
+import { useRef, useState } from 'react';
 
 import type { StepperVisualProps } from '../types';
 
+import { AppPreviewShell, ShellCanvas, ShellSvgLayer } from './AppPreviewShell';
 import {
   BADGE_CUSTOM_BG,
   BADGE_CUSTOM_BORDER,
@@ -15,122 +16,25 @@ import {
   CONNECTIONS,
   type ConnectionDef,
   ENTITIES,
-} from './data/data-model.data';
-import { IconChevronDown } from './icons/data-model-icons';
+} from './data/DataModel.data';
+import { DrawEdge } from './DrawEdge';
+import { IconChevronDown } from './icons/DataModelIcons';
 import {
-  STEPPER_BG,
   STEPPER_BORDER_LIGHT,
   STEPPER_BORDER_MEDIUM,
   STEPPER_BORDER_STRONG,
-  STEPPER_BORDER_SUBTLE,
   STEPPER_CARD_BG,
-  STEPPER_FONT,
-  STEPPER_HEADER_BG,
-  STEPPER_HEADER_BORDER,
-  STEPPER_SHADOW_SM,
   STEPPER_TEXT,
   STEPPER_TEXT_MUTED,
   STEPPER_TEXT_TERTIARY,
 } from './stepper-visual-tokens';
 
-const Wrapper = styled.div`
-  background: ${STEPPER_BG};
-  border-radius: 2px;
-  box-shadow: ${STEPPER_SHADOW_SM};
-  display: flex;
-  flex-direction: column;
-  font-family: ${STEPPER_FONT};
-  height: 92%;
-  margin-left: auto;
-  margin-top: auto;
-  overflow: hidden;
-  width: 88%;
-`;
-
-const Header = styled.div`
-  align-items: center;
-  display: flex;
-  gap: 4px;
-  height: 30px;
-  padding: 0 10px;
-`;
-
-const HeaderLogo = styled.span`
-  align-items: center;
-  background: ${STEPPER_HEADER_BG};
-  border: 1px solid ${STEPPER_HEADER_BORDER};
-  border-radius: 3px;
-  color: ${STEPPER_TEXT};
-  display: flex;
-  height: 14px;
-  justify-content: center;
-  width: 14px;
-`;
-
-const HeaderTitle = styled.span`
-  color: ${STEPPER_TEXT};
-  font-size: 12px;
-  font-weight: 500;
-  line-height: 1.4;
-  padding: 0 2px;
-`;
-
-const HeaderActions = styled.div`
-  align-items: center;
-  display: flex;
-  gap: 8px;
-  margin-left: auto;
-`;
-
-const HeaderBtn = styled.span`
-  align-items: center;
-  border: 1px solid ${STEPPER_BORDER_MEDIUM};
-  border-radius: 4px;
-  color: ${STEPPER_TEXT_MUTED};
-  display: flex;
-  height: 22px;
-  justify-content: center;
-  width: 22px;
-`;
-
-const HeaderCmdBtn = styled.span`
-  align-items: center;
-  border: 1px solid ${STEPPER_BORDER_SUBTLE};
-  border-radius: 4px;
-  color: ${STEPPER_TEXT_TERTIARY};
-  display: flex;
-  font-size: 12px;
-  font-weight: 500;
-  gap: 4px;
-  height: 22px;
-  padding: 0 6px;
-`;
-
-const Canvas = styled.div`
-  background: white;
-  border: 1px solid ${STEPPER_BORDER_MEDIUM};
-  border-radius: 8px;
-  flex: 1;
-  margin: 0 10px 10px;
-  min-height: 0;
-  overflow: hidden;
-  position: relative;
-  user-select: none;
-`;
-
-const SvgLayer = styled.svg`
-  height: 100%;
-  left: 0;
-  pointer-events: none;
-  position: absolute;
-  top: 0;
-  width: 100%;
-`;
-
 const EntityCard = styled.div<{ $hovered: boolean }>`
   backdrop-filter: blur(14px);
   background: ${STEPPER_CARD_BG};
-  border: 1px solid ${({ $hovered }) => ($hovered ? STEPPER_BORDER_STRONG : STEPPER_BORDER_MEDIUM)};
+  border: 1px solid
+    ${({ $hovered }) =>
+      $hovered ? STEPPER_BORDER_STRONG : STEPPER_BORDER_MEDIUM};
   border-radius: 8px;
   box-shadow:
     0 0 2px rgba(0, 0, 0, 0.08),
@@ -295,24 +199,21 @@ export function DataModelVisual({ active }: StepperVisualProps) {
     startY: number;
   } | null>(null);
 
-  const handlePointerDown = useCallback(
-    (entityId: string, event: React.PointerEvent) => {
-      event.preventDefault();
-      (event.currentTarget as HTMLElement).setPointerCapture(event.pointerId);
-      const pos = positions[entityId];
-      dragStartRef.current = {
-        entityId,
-        startX: event.clientX,
-        startY: event.clientY,
-        posX: pos.x,
-        posY: pos.y,
-      };
-      setDragging(entityId);
-    },
-    [positions],
-  );
+  const handlePointerDown = (entityId: string, event: React.PointerEvent) => {
+    event.preventDefault();
+    (event.currentTarget as HTMLElement).setPointerCapture(event.pointerId);
+    const pos = positions[entityId];
+    dragStartRef.current = {
+      entityId,
+      startX: event.clientX,
+      startY: event.clientY,
+      posX: pos.x,
+      posY: pos.y,
+    };
+    setDragging(entityId);
+  };
 
-  const handlePointerMove = useCallback((event: React.PointerEvent) => {
+  const handlePointerMove = (event: React.PointerEvent) => {
     if (!dragStartRef.current) return;
     const { entityId, startX, startY, posX, posY } = dragStartRef.current;
     const dx = event.clientX - startX;
@@ -321,106 +222,35 @@ export function DataModelVisual({ active }: StepperVisualProps) {
       ...prev,
       [entityId]: { x: posX + dx, y: posY + dy },
     }));
-  }, []);
+  };
 
-  const handlePointerUp = useCallback(() => {
+  const handlePointerUp = () => {
     dragStartRef.current = null;
     setDragging(null);
-  }, []);
+  };
 
   const isConnectionHighlighted = (connection: ConnectionDef) =>
     hoveredEntity === connection.from || hoveredEntity === connection.to;
 
   return (
-    <Wrapper style={{ opacity: active ? 1 : 0.7, transition: 'opacity 0.3s' }}>
-      <Header>
-        <HeaderLogo>
-          <svg
-            fill="none"
-            height="9"
-            stroke="currentColor"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth="2"
-            viewBox="0 0 24 24"
-            width="9"
-          >
-            <path d="M10 5a2 2 0 1 0 4 0a2 2 0 0 0 -4 0" />
-            <path d="M19.5 13a2 2 0 1 0 0 4a2 2 0 0 0 0 -4" />
-            <path d="M4.5 13a2 2 0 1 0 0 4a2 2 0 0 0 0 -4" />
-            <path d="M12 7v4" />
-            <path d="M6.5 13l5.5 -2l5.5 2" />
-          </svg>
-        </HeaderLogo>
-        <HeaderTitle>Data model</HeaderTitle>
-        <HeaderActions>
-          <HeaderBtn>
-            <svg fill="none" height="14" stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.5" viewBox="0 0 24 24" width="14">
-              <path d="M6 15l6 -6l6 6" />
-            </svg>
-          </HeaderBtn>
-          <HeaderBtn>
-            <svg fill="none" height="14" stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.5" viewBox="0 0 24 24" width="14">
-              <path d="M6 9l6 6l6 -6" />
-            </svg>
-          </HeaderBtn>
-          <HeaderCmdBtn>
-            <svg fill="none" height="12" stroke="#666" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.5" viewBox="0 0 24 24" width="12">
-              <path d="M12 12m-1 0a1 1 0 1 0 2 0a1 1 0 1 0 -2 0" />
-              <path d="M12 5m-1 0a1 1 0 1 0 2 0a1 1 0 1 0 -2 0" />
-              <path d="M12 19m-1 0a1 1 0 1 0 2 0a1 1 0 1 0 -2 0" />
-            </svg>
-            ⌘K
-          </HeaderCmdBtn>
-        </HeaderActions>
-      </Header>
-
-      <Canvas onPointerMove={handlePointerMove} onPointerUp={handlePointerUp}>
-        <SvgLayer>
-          {CONNECTIONS.map((conn) => {
-            const from = getCardCenter(positions, conn.from);
-            const to = getCardCenter(positions, conn.to);
-            const highlighted = isConnectionHighlighted(conn);
-            const color = highlighted ? STEPPER_BORDER_STRONG : STEPPER_BORDER_MEDIUM;
-
-            const dx = Math.abs(to.x - from.x);
-            const dy = Math.abs(to.y - from.y);
-
-            let pathD: string;
-            let startX = from.x;
-            let startY = from.y;
-            let endX = to.x;
-            let endY = to.y;
-
-            if (dx < 30) {
-              const avgX = (from.x + to.x) / 2;
-              startX = avgX;
-              endX = avgX;
-              pathD = `M${avgX},${from.y} L${avgX},${to.y}`;
-            } else if (dy < 30) {
-              const avgY = (from.y + to.y) / 2;
-              startY = avgY;
-              endY = avgY;
-              pathD = `M${from.x},${avgY} L${to.x},${avgY}`;
-            } else {
-              pathD = `M${from.x},${from.y} L${to.x},${from.y} L${to.x},${to.y}`;
-            }
-
-            return (
-              <g key={`${conn.from}-${conn.to}`}>
-                <path
-                  d={pathD}
-                  fill="none"
-                  stroke={color}
-                  strokeWidth={0.75}
-                  style={{ transition: 'stroke 0.15s' }}
-                />
-                <circle cx={startX} cy={startY} fill={color} r={1.5} />
-                <circle cx={endX} cy={endY} fill={color} r={1.5} />
-              </g>
-            );
-          })}
-        </SvgLayer>
+    <AppPreviewShell active={active} title="Data model">
+      <ShellCanvas
+        onPointerMove={handlePointerMove}
+        onPointerUp={handlePointerUp}
+      >
+        <ShellSvgLayer>
+          {CONNECTIONS.map((conn) => (
+            <DrawEdge
+              key={`${conn.from}-${conn.to}`}
+              circleR={1.5}
+              elbow="horizontal-first"
+              from={getCardCenter(positions, conn.from)}
+              highlighted={isConnectionHighlighted(conn)}
+              id={`${conn.from}-${conn.to}`}
+              to={getCardCenter(positions, conn.to)}
+            />
+          ))}
+        </ShellSvgLayer>
 
         {ENTITIES.map((entity) => {
           const pos = positions[entity.id];
@@ -478,7 +308,7 @@ export function DataModelVisual({ active }: StepperVisualProps) {
             </EntityCard>
           );
         })}
-      </Canvas>
-    </Wrapper>
+      </ShellCanvas>
+    </AppPreviewShell>
   );
 }
