+++ packages/twenty-website-new/src/sections/ProductStepper/visuals/data/layout.data.ts
@@ -1,4 +1,4 @@
-import type { LayoutFieldIconType } from '../icons/layout-icons';
+import type { LayoutFieldIconType } from '../icons/LayoutIcons';
 
 export type FieldDef = {
   icon: LayoutFieldIconType;
